functions/requirements.txt
azure-functions>=1.17.0,<2.0.0
azure-identity>=1.15.0,<2.0.0
azure-keyvault-secrets>=4.7.0,<5.0.0
msal>=1.26.0,<2.0.0
paramiko>=3.4.0,<4.0.0
